Added Admin Activity (#57)
I added a property to the timesheet called AdminActivity that tracks who
approved/rejected a sheet and when. For example, if I approve a sheet,
that property will becomed "Approved by jbrusaw@pdx.edu on <datetime>".
I also updated Timesheet.cshtml to display this property on the "Status"
line if the sheet has been processed. Otherwise it displays "Pending"

The TimesheetContext needs to be updated for this to work

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>