Added Popout Windows on Home Page (#64)
When you left-click a timesheet from the home page, a popup modal
displays containing summary information about the timesheet. You can
also accept or reject a timesheet from this modal. Middle-clicking
or opening the timesheet in a new tab takes you straight to the
details page.

Currently there is no way to check a lock or gain the lock before trying
to accept/reject. This is because the modal is opened without ever
calling a controller, and it would take an excessive amount of C# code
in the view to accomplish this, which is considered bad practice. We'll
need to adress the best way to handle this later

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>