Massive Database Refactor (#72)
* Massive Database Refactor

Refactored the database to fit our final schema. I had to make a lot of
changes to the code to fit this new schema, including renaming
"Timesheet" in a lot of place to "Submission" so that in the future we
can accomodate mileage forms as well as timesheets.

The dbcontext comes with 3 preloaded indices: context.Submissions for
every submission, context.Timesheets for just the timesheets, and
context.MileageForms for just the mileage forms.

I couldn't get eager loading to work, so relational databases (such as
the Lock table or the TimeEntry tables) need to be explicity loaded. See
the Timesheet() function in HomeController for an example. Lazy loading
may be possible in a future update, but ironically that was too much
effort for this one.

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>

* Moved Models to Common

Takes a little work, but it's not too bad.

Signed-off-by: David Post <post@pdx.edu>