Textract object (#73)
* Adding in Textract Document Format

This format will allow us to convert a response into a structured
data format that is traversible and makes sense. As an added
benefit, it can be serialized and deserialized from the Json
response given by AWS.

Signed-off-by: David Post <post@pdx.edu>

* Completed first stage digitization of textract

This takes textract json and digitizes it into a rather simple
set of objects. The next stage will be to construct the structure
of the data based off the information given so far.

Signed-off-by: David Post <post@pdx.edu>

* Added words to lines and cells to tables

Along with adding lines and tables to pages. This is the beginning
of the structured data.

Signed-off-by: David Post <post@pdx.edu>

* Got tables printing out cell information

It creates an internal structure linking the cells together in a
table format and is able to format it nicely for string output.

Signed-off-by: David Post <post@pdx.edu>

* Added confidence and page access methods

We can now access the confidence of each block and which page
it belongs to.

Signed-off-by: David Post <post@pdx.edu>

* Added Textract Response Parser

Now we can parse also from the textract response which works by simply
converting it to Json then passing that to the json parser.

Signed-off-by: David Post <post@pdx.edu>

* Got Textract Working in C#

Sadly it doesn't handle PDF's from a bytestream, but we do have images
working. We'll need to setup a S3 bucket for handling PDF files.

Signed-off-by: David Post <post@pdx.edu>

* Removed Textract PDF processing

This removes the PDF part of it for now. We need to add in S3 bucket
to have this part working as for whatever reason AWS accepts byte
string for images but not for PDFs.

Signed-off-by: David Post <post@pdx.edu>

* Removed unnecessary assemblers

A lot of weird dependencies somehow crept up, but are not used.

Signed-off-by: David Post <post@pdx.edu>

* Added support for key-value pairs

This is rudamentary support at best right now. As we begin to use it
more features may be added.

Signed-off-by: David Post <post@pdx.edu>

* Added Summary data for lines and tables

To be complete it's now printing summary data for all entities.

Signed-off-by: David Post <post@pdx.edu>