Added Handling of Multiple Image Uploads (#90)
* Added Test for Merging Documents

We now have a test for merging documents. Next step is to get them to
merge.

* Added Merging of TextractDocuments

Now we can merge multiple TextractDocuments together. This sets
each page inside the merged portion in the same order but with
the page number being counted from the end of the original
document.

* Added Handling of Multiple Pages

Now we can merge multiple pages into the same document.

Signed-off-by: David Post <post@pdx.edu>