Mileage Forms, Lock Column, Mass Download PDFs, Pay Periods, and Various Bug Fixes (#95)
* Added Support for Mileage Forms

I made a lot of changes to a lot of files, and now the AdminUI supports
Mileage Forms as well as Timesheet Forms

Note this will require an Update-Database -Context SubmissionContext
command to function correctly. I'd also recommend deleting your database
entirely and re-running both Update-Database commands so that the
MockData populates with some sample Mileage Forms

* Added Lock Column

I added a column to the main page to show who has the current lock so
it's easier to identify at a glance which sheets can be worked on. I
skimped on the filter because it would require a lot of effort for
minimal return on usability and we have more important things to do

* Added Mass Download PDF function

Added a button on the main page to download every PDF matching the
selected criteria

As of now, it's very inefficent. It takes ~2 seconds to generate a PDF,
of which the vast majority of that time is spent adding the images of
the timesheets. For 1,600 sheets, that means it'll take ~45 minutes to
render and download them all. I spent about an hour researching
improvements, but all I could come up with is spending money to get a
better PDF Renderer : /

I highly recommend filtering your sheets to 2-5 before attempting to
download. Otherwise you'll be waiting awhile

* Added Pay Periods

Added Pay Periods and gave Administrators the ability to create,
update, delete, and start pay periods. There is a global
"CurrentPayPeriod" variable that determines the default to/from dates
when filtering sheets, as well as a warning that displays to all users
when running low on dates.

This required another Database Context, so please run "Update-Database
-Context PayPeriodContext" to enable functionality

* Added Pay Periods

Added Pay Periods and gave Administrators the ability to create,
update, delete, and start pay periods. There is a global
"CurrentPayPeriod" variable that determines the default to/from dates
when filtering sheets, as well as a warning that displays to all users
when running low on dates.

This required another Database Context, so please run "Update-Database
-Context PayPeriodContext" to enable functionality

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>

* Fixed and added tests for payperiod context so itll all pass.

* Added the Update PayPeriodContext to the Readme.

Signed-off-by: Larry Chiem <clarry@pdx.edu>

* Very small changes to Index

Forgot to add "required" fields to MileageForm, and changed input type
on ProviderId from 'select' to 'text'

* Various bug fixes

Fixed a few database concurrency issues, a date sort issue, and replaced
the URIs with updated onces to fix broken links in the old timesheets

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>

Co-authored-by: Larry Chiem <clarry@pdx.edu>