Added Better Indexing Algorithm (#139)
* Added Levenshtein Distance function

Got the distance function to work and tests added for it.

* Completed NGLD function

The NGLD function is complete and working well. Next step is finishing
up the string matching.

* Got extra and missing inputs tests passing

We can now pass in two lists and have it find matches. If the second
input is less than the first then it removes the worse matching
duplicates.

A possible optimization would be to reassign it to the next best and
recurse over that.

* Got the string matching with tolerance of 0.31

This allows for testing for front page if it contains the characteristic
that there is an element within 0.31 NGLD of "Service Provided On:".

* Changed to only update the provider date once

There are actually 3 occurrences of "Date" on the back and we only want
to update it with the second date and not the third.

* Removed commented out code

The provider sign date is taken care of earlier and this line of code
is no longer needed. Replaced it with a comment directing attention
to the place where it happens.

Signed-off-by: David Post <post@pdx.edu>