Fixed Selection Item and More Front Page Detection (#147)
* Fixed selection item exception

In some cases there are no relations and it throws an error. We simply
catch this and move on.

Also fixes an error in the AbstractFormObject where an empty string key
is already added.

* Tightened tolerance on front page detection

Tolerance was too high and we were hitting the back page sometimes. This
tolerance should allow still for four edits but hit fewer false
positives.

* Added guard against false positive on front page

To allow more edits we need to loosen the tolerance, but to protect
against false positive we want to double check that this isn't the back
page. So we verify that it has something that should be only on the
front page, and then check that it doesn't contain something only on the
back page.

* Corrected a comment and removed uneeded using

A little cleanup of the code.

* Changed Approval to String

This should be the initials of the submitter, but this will be hard to
detect. Instead this will be a blank string from Appserver. In PWA
Submission we will accept a string back.

* Updated README.md

Cleared up the updating of the database to indicate that the
submissioncontext needs to be loaded from the AdminUI project.

* Improved Front Page Detection

There are various versions of the forms with different front and back
pages. Trying to find a common element in all has been a challenge.

Added custom substring function to take a substring up to the end of
the line.

Signed-off-by: David Post <post@pdx.edu>