Improved query functionality (#152)
* Refactored the Views

* Home now has a single Index with partial views to render a
  MileageFormTable and a TimesheetTable
* Submission now has a single Index with partial views to render a
  MileageFormTable and a TimesheetTable
* Pulled the navbar out of _LayoutContainer and _LayoutNoContainer and
  created a partial view out of that

* Refactored File Structure

Moved PayPeriod.cs to Models and MockData.cs to Data

* Refactored Controller Logic

* Refactored ToPDF() function so Submission.cs handles the majority of
  work and delegates the individual entry tables to its children
  Timesheet.cs and MileageForm.cs
* Combined ModalProcess() and Process() into one function that returns
  different views depending on where it was called from

* Improved Query Functionality

Previously, each function got all the timesheets from the database and
filtered the list down 1 filter at a time.

Now, a single function builds the query up and queries the database
once, which should be significantly faster. This is also better for
scalability and code maintenance down the road.

I also renamed a couple variables in HomeController to make Curtis
happy. There are a lot more variables that need improvement, but that's
another ticket for another time

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>