Synchronized Staging and Submission Ids (#163)
* Synchronizing the staging and submission ids

It's a little more complicated than I would like, but it's workable in
normal situations. When running through our test endpoints we do run
into the issue of overlapping submissions, but in practice that should
not be the case, and we can also change the way we handle this in
testing to accommodate the normal workflow by having SubmitTest take
the staging we want to submit and copy it in the staging table.

* Synchronize Stagings and Submissions Index

To keep the id's synchronized we reseed the index with the max of
stagings and submissions. For SubmitTest we take the old submission
and make a new one to avoid trying to insert one that is already there.

Signed-off-by: David Post <post@pdx.edu>